gerenciador_leads/requirements.txt
# Dependências do Gerenciador de Leads
flask>=2.3.0
pandas>=2.0.0
openpyxl>=3.1.0
requests>=2.31.0
pywebview>=4.4.0
pyinstaller>=6.0.0
